refactor(cli): split cli in actions files
